Updates
two news files, Logic, add transitions of pmv_low and High.

+++ Logic.h
@@ -5,7 +5,7 @@
 #include <Keypad.h>
 #include "states.h"
 //Externse
-extern int alarmCount;
+extern byte alarmCount;
 extern byte indx;
 extern char key;
 extern byte inputIndex;
